requirements.txt
anthropic
fastapi
uvicorn
asyncio
pytest
black
flake8
python-dotenv
EOF < /dev/null